requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via starlette
cachetools==6.2.2
    # via google-auth
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
docstring-parser==0.17.0
    # via google-cloud-aiplatform
fastapi==0.115.5
    # via -r requirements.in
google-api-core[grpc]==2.28.1
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.43.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   google-genai
google-cloud-aiplatform==1.56.0
    # via -r requirements.in
google-cloud-bigquery==3.38.0
    # via google-cloud-aiplatform
google-cloud-core==2.5.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.16.0
    # via -r requirements.in
google-cloud-resource-manager==1.15.0
    # via google-cloud-aiplatform
google-cloud-storage==2.18.2
    # via
    #   -r requirements.in
    #   google-cloud-aiplatform
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==0.6.0
    # via -r requirements.in
google-resumable-media==2.8.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.72.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.3
    # via google-cloud-resource-manager
grpcio==1.76.0
    # via
    #   google-api-core
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
numpy==2.3.5
    # via shapely
packaging==25.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
pillow==11.3.0
    # via google-genai
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-firestore
    #   google-cloud-resource-manager
protobuf==4.25.8
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-firestore
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
pydantic-core==2.23.4
    # via pydantic
pypdf==4.3.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
python-multipart==0.0.20
    # via -r requirements.in
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
rsa==4.9.1
    # via google-auth
shapely==2.1.2
    # via google-cloud-aiplatform
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   grpcio
    #   pydantic
    #   pydantic-core
urllib3==2.5.0
    # via requests
uvicorn==0.30.6
    # via -r requirements.in
websockets==14.2
    # via google-genai
